requirements.txt
alembic==1.9.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
flask-marshmallow==0.14.0
Flask-Migrate==4.0.3
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.0
psycopg2==2.9.5
python-dotenv==0.21.1
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.0
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
WTForms==3.0.1
